refactor: move Design tab FAB into the Widgets subtab only
The FAB previously overlaid all three Design subtabs via a page-change
visibility callback in TabDesignFragment. Moving the FAB view itself into
SubtabWidgetsFragment's own layout is simpler: it's naturally only present
when that subtab's view exists, no visibility bookkeeping needed.
