+++ app/src/main/res/layout/fragment_tab_design.xml
@@ -37,15 +37,5 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent" />
 
-        <dev.oneuiproject.oneui.widget.ScrollAwareFloatingActionButton
-            android:id="@+id/fabDesign"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="end|bottom"
-            android:layout_margin="24dp"
-            android:src="@drawable/ic_oui_info_outline"
-            app:fabSize="normal"
-            app:layout_behavior="@null" />
-
     </dev.oneuiproject.oneui.widget.RoundedFrameLayout>
 </LinearLayout>
